500,000 takes about 6 seconds to execute and I believe 50,000 is more reasonable and gives me the confidence I need that it works as intended.
If a user made 10 entries a day for an entire year, that's only about 3600, so 50,000 is more than plenty for a user's entire lifetime using the app (more than likely).

For reference 10 entries per day per year is 3650. For 10 years that's 36,500. So 50,000 for one user should be great and there's really no reason to belive it wouldn't keep working after that.

I also don't care about collision between users because entries are namespaced by user id.