Add data visualization dashboard (#39)
* Initial configuration of library with routes to module and component

* Refactor getting sub-nav router into function

* Fix routing links

Having the pointer-events: none on the anchor actually disabled routing at all, this means I had to add it to the child elements individually

This change also reintroduced the issue where clicking on just the color of the nav option did not trigger an tracking event, so the solution was to make the anchor the full size of the button

* Remove better code hub config

It got readded when I merged master into develop to handle merge conflicts...

* Update README

* Initial work into getting dashboard up and running

* Restructure tracker selectors to prepare to pass data to dashboard
* Add graph service to handle aggregating data for visualizations

* Refactor some tracker reducer logic into a helper in hopes that it makes generating data for the other planned graphs easier

* Displaying the time vs date graph

This will later move to it's own component and the inputs will be refactored to be more reusable...this is just first steps

* Displaying all desired graph data as it is being selected right now

These obviously need to be improved upon, but this is a good starting point

* Added a little color scheme utility to get themes bundled with the library because most of them are actually pretty good

* add date-fns library

* Using advanced pie chart

* Large refactoring of application structure

* Remove TrackerApp module that was previously rendering the app routes in a child router-outlet
* I was unable to directly link to a child router-outlet when trying to get the returnUrl to work, so I restructured to not have my app work that way
* This inadvertently cleaned up how the routing works anyway...the url is no longer annoyingly cluttered by the child router outlet info
* Had to restructure some CSS stuff to work with new changes...went ahead and made some minor tweaks to the components that had been bugging me

* Update packages

* Added spinner

Still need to figure out how I want to this stuff to be organized

* Consolidated all routing login into the root routing module

I had theme separate before, but that was when I actually had a feature module in charge of routing of some kind...now that that's not a thing I feel it makes more sense to have all the routing logic in one place...at least until it makes more sense to split it up again

* Add unit tests for date utils

* Move router utils to root reducer, refactor graph service, and add some more tests

* test generating unique ids for a user

* add firefox test launcher

* Added serious and not so serious headless test runners

* 500,000 takes about 6 seconds to execute and I believe 50,000 is more reasonable and gives me the confidence I need that it works as intended.

If a user made 10 entries a day for an entire year, that's only about 3600, so 50,000 is more than plenty for a user's entire lifetime using the app (more than likely).

For reference 10 entries per day per year is 3650. For 10 years that's 36,500. So 50,000 for one user should be great and there's really no reason to belive it wouldn't keep working after that.

I also don't care about collision between users because entries are namespaced by user id.

* Refactor and add some more tests

* It seems like running this test periodically messes things up.

Need to investigate, but for now I will just skip it. It has mostly served it's purpose and that code will not ever change in the future.

* Add more tests and refactor some

* Modified the UUID test to be synchronous

* All utils are tested!

* Added a lot of tests for reducers...nearly done

* Add more tests and refactor some by adding a total time pipe

* All reducers and selectors tested

* small style fixes

* added default value for graphs

when history entries don't have a platform or game, their corresponding graphs don't have a good label...default value fixes this

* add function to get the week of a day

* Graphs are now displaying data for the current calendar week starting on sunday

* Utilize another date fns function

* Completed tests for filterGraphDataByDateRange

* add travis config for running tests

* trying to get tests running on travis

* attempt #3

* fix tests

* add all tests for new graph helpers

* displaying total time and a basic custom tooltip

* change default value for empty graph key

* make tests for tool tip pass

* add some tooltip styles

* Finish tool tip styling

* Working no dashboard layout

* add dashboard styling